dv-export/python/requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
SQLAlchemy==2.0.29
visier_connector==0.9.17
